update CustomerServiceImpl (validation); update ExcursionServiceImplTest

+++ src/main/java/cz/muni/fi/pa165/jtravelagency/service/CustomerServiceImpl.java
@@ -26,23 +26,38 @@ public class CustomerServiceImpl implements CustomerService {
     private CustomerDAO customerDao;
 
     public void create(CustomerDTO customerDto) {
+        validateCustomer(customerDto);
+        if (customerDto.getId() != null) {
+            throw new IllegalArgumentException("Customer's id is not null.");
+        }
         Customer customer = DTOAndEntityMapper.dtoToEntity(customerDto, Customer.class);
         customerDao.createCustomer(customer);
         customerDto.setId(customer.getId());
     }
 
     public CustomerDTO get(Long id) {
+        if (id == null) {
+            throw new IllegalArgumentException("Id cannot be null.");
+        }
         Customer customer = customerDao.getCustomer(id);
         CustomerDTO customerDto = DTOAndEntityMapper.entityToDto(customer, CustomerDTO.class);
         return customerDto;
     }
 
     public void update(CustomerDTO customerDto) {
+        validateCustomer(customerDto);
+        if (customerDto.getId() == null) {
+            throw new IllegalArgumentException("Id cannot be null.");
+        }
         Customer customer = DTOAndEntityMapper.dtoToEntity(customerDto, Customer.class);
         customerDao.updateCustomer(customer);
     }
 
     public void delete(CustomerDTO customerDto) {
+        validateCustomer(customerDto);
+        if (customerDto.getId() == null) {
+            throw new IllegalArgumentException("Id cannot be null.");
+        }
         Customer customer = DTOAndEntityMapper.dtoToEntity(customerDto, Customer.class);
         customerDao.deleteCustomer(customer);
     }
@@ -60,4 +75,15 @@ public class CustomerServiceImpl implements CustomerService {
         customerDao.setDeletedStatus(customer);
     }
 
+    private void validateCustomer(CustomerDTO customerDto) {
+        if (customerDto == null) {
+            throw new IllegalArgumentException("Customer cannot be null.");
+        }
+        if (customerDto.getFirstName() == null) {
+            throw new IllegalArgumentException("Customer's first name cannot be null.");
+        }
+        if (customerDto.getLastName() == null) {
+            throw new IllegalArgumentException("Customer's last name cannot be null.");
+        }
+    }
 }
